+++ app/src/main/java/com/elliottsoftware/calftracker/data/sources/implementations/FireBaseAuthentication.kt
@@ -95,4 +95,9 @@ class FireBaseAuthentication : AuthenticationSource {
         awaitClose()
     }
 
+    override fun currentUserEmail():String{
+        return auth.currentUser?.email!!
+
+    }
+
 }
