+++ app/src/main/java/com/elliottsoftware/calftracker/data/sources/implementations/FireBaseFireStore.kt
@@ -2,12 +2,14 @@ package com.elliottsoftware.calftracker.data.sources.implementations
 
 import com.elliottsoftware.calftracker.data.sources.DatabaseSource
 import com.elliottsoftware.calftracker.domain.models.Response
+import com.elliottsoftware.calftracker.domain.models.fireBase.FireBaseCalf
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
+import timber.log.Timber
 
 class FireBaseFireStore:DatabaseSource {
     private val db: FirebaseFirestore = Firebase.firestore
@@ -27,6 +29,28 @@ class FireBaseFireStore:DatabaseSource {
                     e ->
                 trySend(Response.Failure(Exception(" Error! Please try again")))
             }
+
+        awaitClose()
+    }
+
+    override fun createCalf(calf: FireBaseCalf, userEmail: String): Flow<Response<Boolean>> = callbackFlow {
+        trySend(Response.Loading)
+
+            val collection = db.collection("users").document(userEmail)
+                .collection("calves")
+            val document = collection.document()
+            val id = document.id
+            calf.id = id
+
+            collection.document(id).set(calf)
+                .addOnSuccessListener { document ->
+                    Timber.d( "DocumentSnapshot written with ID: ${calf.id}")
+                    trySend(Response.Success(true))
+                }
+                .addOnFailureListener { e ->
+                    Timber.e( "Error adding document", e)
+                    trySend(Response.Failure(e))
+                }
 
         awaitClose()
     }
